+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 1.  [#2265](https://github.com/influxdata/chronograf/pull/2265): Autofocus dashboard query editor
 1.  [#4429](https://github.com/influxdata/chronograf/pull/4429): Fix query editor flickering on update
 1.  [#4452](https://github.com/influxdata/chronograf/pull/4452): Improve log search spinner info
+1.  [#4525](https://github.com/influxdata/chronograf/pull/4525): Move expanded log message copy button to top right
 
 ### UI Improvements
 1.  [#4236](https://github.com/influxdata/chronograf/pull/4236): Add spinner when loading logs table rows